requirements.txt
Django==5.2.1 
djangorestframework 
psycopg2-binary 
python-dotenv